Handle timeout related to materialized views on HiveServer2 startup
Today, the default behavior of HiveServer2 is to load all materialized
views from the metastore, then to load them in a registry.

In this change, I add the possibility to simply skip this step.
I also modify the parallelism when looking up views: instead of
retrieving all dbs, then all tables for each db and load it in the registry asynchronously,
we look up tables on a separate thread to allow for a faster startup
of HiveServer2.